Revert "feat(ui): share shell providers with module federation remotes"
