+++ ui/package.json
@@ -14,17 +14,16 @@
   "dependencies": {
     "@stomp/stompjs": "^7.1.1",
     "@tanstack/react-query": "^5.85.9",
-    "@xyflow/react": "^12.0.2",
     "lucide-react": "^0.542.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "@xyflow/react": "^12.0.2",
     "react-resizable-panels": "^3.0.5",
     "react-router-dom": "^7.8.2",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
-    "@originjs/vite-plugin-federation": "^1.4.1",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.6.1",
