--- ui/src/lib/theme.tsx
@@ -1,36 +1,0 @@
-import { ReactNode, createContext, useContext, useEffect, useMemo, useState } from 'react'
-
-export type Theme = 'dark' | 'light'
-
-type ThemeContextValue = {
-  theme: Theme
-  setTheme: (theme: Theme) => void
-}
-
-const ThemeContext = createContext<ThemeContextValue>({
-  theme: 'dark',
-  setTheme: () => undefined,
-})
-
-export type ThemeProviderProps = {
-  children: ReactNode
-  initialTheme?: Theme
-}
-
-export function ThemeProvider({ children, initialTheme = 'dark' }: ThemeProviderProps) {
-  const [theme, setTheme] = useState<Theme>(initialTheme)
-
-  useEffect(() => {
-    if (typeof document !== 'undefined') {
-      document.documentElement.dataset.theme = theme
-    }
-  }, [theme])
-
-  const value = useMemo(() => ({ theme, setTheme }), [theme])
-
-  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
-}
-
-export function useTheme() {
-  return useContext(ThemeContext)
-}
