+++ ui/src/main.tsx
@@ -1,10 +1,18 @@
+import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { BrowserRouter } from 'react-router-dom'
 import './styles/globals.css'
 import App from './App.tsx'
-import ShellProviders from './ShellProviders'
+
+const queryClient = new QueryClient()
 
 createRoot(document.getElementById('root')!).render(
-  <ShellProviders>
-    <App />
-  </ShellProviders>,
+  <StrictMode>
+    <QueryClientProvider client={queryClient}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </QueryClientProvider>
+  </StrictMode>,
 )
