--- ui/src/shared/index.ts
@@ -1,14 +1,0 @@
-export {
-  ConfigProvider,
-  getConfig,
-  resetConfig,
-  setConfig,
-  subscribeConfig,
-  useConfig,
-} from '../lib/config'
-export { ShellProviders } from '../ShellProviders'
-export { ThemeProvider, useTheme } from '../lib/theme'
-export type { Theme } from '../lib/theme'
-export { useUIStore } from '../store'
-export { createShellRoot } from '../remote/createShellRoot'
-export type { ShellRoot, ShellRootOptions } from '../remote/createShellRoot'
